requirements.txt
langchain
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
groq
cassio
beautifulsoup4
langchain-groq
langchainhub
sentence_transformers
PyPDF2
langchain-objectbox
opentelemetry-api==1.13.0
opentelemetry-sdk==1.13.0
protobuf==3.20.3